last changes before pull request
