+++ app/src/main/java/com/crypto/katip/HomeActivity.java
@@ -19,8 +19,9 @@ import com.crypto.katip.login.LoginRepository;
 import com.crypto.katip.database.DbHelper;
 import com.crypto.katip.models.Chat;
 import com.crypto.katip.models.LoggedInUser;
-import com.crypto.katip.viewmodels.home.HomeViewModel;
-import com.crypto.katip.viewmodels.home.HomeViewModelFactory;
+import com.crypto.katip.ui.home.ChatAdderFragment;
+import com.crypto.katip.ui.home.HomeViewModel;
+import com.crypto.katip.ui.home.HomeViewModelFactory;
 
 import java.util.ArrayList;
