+++ app/src/main/java/com/crypto/katip/ui/home/HomeViewModel.java
@@ -1,4 +1,4 @@
-package com.crypto.katip.viewmodels.home;
+package com.crypto.katip.ui.home;
 
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
