+++ app/src/main/java/com/crypto/katip/ui/home/HomeViewModelFactory.java
@@ -1,16 +1,16 @@
-package com.crypto.katip.viewmodels.register;
+package com.crypto.katip.ui.home;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-public class RegisterViewModelFactory implements ViewModelProvider.Factory {
+public class HomeViewModelFactory implements ViewModelProvider.Factory {
     @NonNull
     @Override
     @SuppressWarnings("unchecked")
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass.isAssignableFrom(RegisterViewModel.class)) {
-            return (T) new RegisterViewModel();
+        if (modelClass.isAssignableFrom(HomeViewModel.class)) {
+            return (T) new HomeViewModel();
         } else {
             throw new IllegalArgumentException("Unknown ViewModel class");
         }
