+++ app/src/main/java/com/crypto/katip/ui/login/LoginViewModelFactory.java
@@ -1,4 +1,4 @@
-package com.crypto.katip.viewmodels.login;
+package com.crypto.katip.ui.login;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
