+++ app/src/main/java/com/crypto/katip/ui/register/RegisterResult.java
@@ -1,4 +1,4 @@
-package com.crypto.katip.viewmodels.register;
+package com.crypto.katip.ui.register;
 
 import com.crypto.katip.models.User;
