+++ app/src/main/java/com/crypto/katip/ui/register/RegisterViewModel.java
@@ -1,4 +1,4 @@
-package com.crypto.katip.viewmodels.register;
+package com.crypto.katip.ui.register;
 
 import android.content.Context;
