+++ app/src/main/java/com/crypto/katip/ui/register/RegisterViewModelFactory.java
@@ -1,16 +1,16 @@
-package com.crypto.katip.viewmodels.home;
+package com.crypto.katip.ui.register;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-public class HomeViewModelFactory implements ViewModelProvider.Factory {
+public class RegisterViewModelFactory implements ViewModelProvider.Factory {
     @NonNull
     @Override
     @SuppressWarnings("unchecked")
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass.isAssignableFrom(HomeViewModel.class)) {
-            return (T) new HomeViewModel();
+        if (modelClass.isAssignableFrom(RegisterViewModel.class)) {
+            return (T) new RegisterViewModel();
         } else {
             throw new IllegalArgumentException("Unknown ViewModel class");
         }
